requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
boto3==1.34.78
botocore==1.34.78
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
decord==0.6.0
distlib==0.3.8
easydict==1.13
elastic-transport==8.13.0
elasticsearch==8.13.1
et-xmlfile==1.1.0
eva-decord==0.6.1
exceptiongroup==1.2.0
filelock==3.13.3
Flask==3.0.3
fonttools==4.50.0
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.43.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.63.0
httplib2==0.22.0
huggingface-hub==0.22.2
identify==2.5.36
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdit-py-plugins==0.4.0
mdurl==0.1.2
memray==1.11.0
mocker==1.1.1
mpmath==1.3.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pip==23.3.1
platformdirs==4.2.1
plotly==5.20.0
pluggy==1.4.0
pre-commit==3.7.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
scipy==1.13.0
SecretStorage==3.3.3
setuptools==59.6.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
ssave-ai-packages==0.2.4
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
textual==0.55.1
threadpoolctl==3.5.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==1.26.18
versioneer==0.29
virtualenv==20.26.1
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.37.1
zipp==3.18.1
